Keyboard Update
Keyboard is working with backspaces.
Commands aren’t working.

+++ terminal.hpp
@@ -63,7 +63,7 @@ private:
     
     void clearLine(uint8_t from, uint8_t to);
     
-    void scrollUp(uint8_t lineNum);
+    void scrollUp(uint64_t lineNum);
     
 public:
     Terminal();
